rebasing use-yaml-config branch
