+++ src/CGx/utils/generate_square_mesh.py
