+++ src/CGx/utils/plot_scaling.py
